use-cases/task-manager/python/requirements (1).txt
fastapi
uvicorn
pydantic
httpx
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
python-multipart
